[algorithms] Break include cycles caused by GC support in covered_by and within
