+++ include/boost/geometry/algorithms/detail/within/implementation.hpp
@@ -30,7 +30,6 @@
 #include <boost/geometry/algorithms/detail/within/multi_point.hpp>
 #include <boost/geometry/algorithms/detail/within/point_in_geometry.hpp>
 #include <boost/geometry/algorithms/detail/relate/implementation.hpp>
-#include <boost/geometry/algorithms/detail/relate/implementation_gc.hpp>
 #include <boost/geometry/algorithms/detail/relate/relate_impl.hpp>
 
 #include <boost/geometry/core/access.hpp>
@@ -292,23 +291,6 @@ struct within<MultiPolygon, Polygon, multi_polygon_tag, polygon_tag>
 template <typename MultiPolygon1, typename MultiPolygon2>
 struct within<MultiPolygon1, MultiPolygon2, multi_polygon_tag, multi_polygon_tag>
     : public detail::within::use_relate
-{};
-
-// GC
-
-template <typename Geometry1, typename Geometry2>
-struct within<Geometry1, Geometry2, geometry_collection_tag, geometry_collection_tag>
-    : detail::within::use_relate
-{};
-
-template <typename Geometry1, typename Geometry2, typename Tag1>
-struct within<Geometry1, Geometry2, Tag1, geometry_collection_tag>
-    : detail::within::use_relate
-{};
-
-template <typename Geometry1, typename Geometry2, typename Tag2>
-struct within<Geometry1, Geometry2, geometry_collection_tag, Tag2>
-    : detail::within::use_relate
 {};
 
 } // namespace dispatch
